Fix ThowNativeErr typo
