--- packages/react-native/examples/expo-app/modules/thow-err-module/android/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<manifest package="expo.modules.thowerrmodule">
-</manifest>
