--- packages/react-native/examples/expo-app/modules/thow-err-module/android/src/main/java/expo/modules/thowerrmodule/ThowErrModuleView.kt
@@ -1,7 +1,0 @@
-package expo.modules.thowerrmodule
-
-import android.content.Context
-import expo.modules.kotlin.AppContext
-import expo.modules.kotlin.views.ExpoView
-
-class ThowErrModuleView(context: Context, appContext: AppContext) : ExpoView(context, appContext)
