--- packages/react-native/examples/expo-app/modules/thow-err-module/expo-module.config.json
@@ -1,9 +1,0 @@
-{
-  "platforms": ["ios", "android", "web"],
-  "ios": {
-    "modules": ["ThowErrModule"]
-  },
-  "android": {
-    "modules": ["expo.modules.thowerrmodule.ThowErrModule"]
-  }
-}
