--- packages/react-native/examples/expo-app/modules/thow-err-module/src/ThowErrModuleView.web.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-
-import { ThowErrModuleViewProps } from './ThowErrModule.types';
-
-export default function ThowErrModuleView(props: ThowErrModuleViewProps) {
-  return (
-    <div>
-      <span>{props.name}</span>
-    </div>
-  );
-}
