+++ packages/react-native/examples/expo-app/modules/throw-err-module/android/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest package="expo.modules.throwerrmodule">
+</manifest>
