+++ packages/react-native/examples/expo-app/modules/throw-err-module/android/src/main/java/expo/modules/thowerrmodule/ThrowErrModuleView.kt
@@ -1,0 +1,7 @@
+package expo.modules.throwerrmodule
+
+import android.content.Context
+import expo.modules.kotlin.AppContext
+import expo.modules.kotlin.views.ExpoView
+
+class ThrowErrModuleView(context: Context, appContext: AppContext) : ExpoView(context, appContext)
