+++ packages/react-native/examples/expo-app/modules/throw-err-module/ios/ThrowErrModule.podspec
@@ -1,5 +1,5 @@
 Pod::Spec.new do |s|
-  s.name           = 'ThowErrModule'
+  s.name           = 'ThrowErrModule'
   s.version        = '1.0.0'
   s.summary        = 'A sample project summary'
   s.description    = 'A sample project description'
