+++ packages/react-native/examples/expo-app/modules/throw-err-module/ios/ThrowErrModuleView.swift
@@ -2,6 +2,6 @@ import ExpoModulesCore
 
 // This view will be used as a native component. Make sure to inherit from `ExpoView`
 // to apply the proper styling (e.g. border radius and shadows).
-class ThowErrModuleView: ExpoView {
+class ThrowErrModuleView: ExpoView {
   
 }
