+++ packages/react-native/examples/expo-app/modules/throw-err-module/src/ThrowErrModule.ts
@@ -2,4 +2,4 @@ import { requireNativeModule } from 'expo-modules-core';
 
 // It loads the native module object from the JSI or falls back to
 // the bridge module (from NativeModulesProxy) if the remote debugger is on.
-export default requireNativeModule('ThowErrModule');
+export default requireNativeModule('ThrowErrModule');
