+++ packages/react-native/examples/expo-app/modules/throw-err-module/src/ThrowErrModule.web.ts
