+++ packages/react-native/examples/expo-app/modules/throw-err-module/src/ThrowErrModuleView.web.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+import { ThrowErrModuleViewProps } from './ThrowErrModule.types';
+
+export default function ThrowErrModuleView(props: ThrowErrModuleViewProps) {
+  return (
+    <div>
+      <span>{props.name}</span>
+    </div>
+  );
+}
